% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_stbenchi_db}
\alias{connect_stbenchi_db}
\title{Connect to ST_Benchi Database}
\usage{
connect_stbenchi_db(file, readonly = T)
}
\arguments{
\item{file}{the database file}

\item{readonly}{whether the file shall be opened read-only}
}
\value{
a database connection
}
\description{
This function provides an easy way of connecting to the ST_Benchi results database.
}
\examples{
connect_stbenchi_db("results.db")
}
