% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_1024.R
\name{convert_to_bytes_1024}
\alias{convert_to_bytes_1024}
\title{Convert prefixed byte string to bytes, conversion factor 1024}
\usage{
convert_to_bytes_1024(values)
}
\arguments{
\item{values}{the size in bytes containing a unit}
}
\value{
the size in bytes
}
\description{
This function converts a string containing value and a byte size unit into the
corresponding raw byte value. It is assumed that 1024 byte = 1 kibibyte etc.
}
\details{
Function adapted from https://www.garysieling.com/blog/r-convert-from-gib-mib-kib-to-bytes/
}
\examples{
convert_to_bytes_1024("10 MiB")
convert_to_bytes_1024("2KiB")
convert_to_bytes_1024("42 B")
}
