% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{create_stage_graph}
\alias{create_stage_graph}
\title{Plot Resource Utilization Graph for One Benchmark Run}
\usage{
create_stage_graph(conn, run_id_input, bench_param, stage_names)
}
\arguments{
\item{conn}{the database connection}

\item{run_id_input}{the run id to plot}

\item{bench_param}{a character vector of at most 2 stats that shall be plotted}

\item{stage_names}{a character vector of stages to be plotted}
}
\value{
a ggplot2 plot of the two selected resource utilization metrics
}
\description{
This function is a wrapper around the @seealso [plot_docker_stats_single()] function,
which creates a single resource utilization graph for at most two variables.
}
\examples{
# for run with id 29, create the resource utilization graph for CPU and Memory usage
create_stage_graph(conn, 29, c("CPUPerc", "MemUsage"), c("ingestion"))
}
