% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{generate_merged_factor}
\alias{generate_merged_factor}
\title{Generate Factor for Combination of Parameters}
\usage{
generate_merged_factor(df, unique_params)
}
\arguments{
\item{df}{the parsed dataframe}

\item{unique_params}{the parameters}
}
\value{
a character vector as an input for the factor
}
\description{
This helper function takes all values of a well-defined set of parameters and
creates the cross product of their unique values, concatenates each combination into
a single string and returns the resulting character vector of strings.
}
\examples{
generate_merged_factor(df, c("system", "vectorize_type"))
generate_merged_factor(df, c("system", "recipe", "recipe"))
}
