% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_timings.R
\name{get_run_timings}
\alias{get_run_timings}
\title{Get Timings and Parameter Information for a Certain Run}
\usage{
get_run_timings(conn, run_id_input)
}
\arguments{
\item{conn}{the database connection}

\item{run_id_input}{the id of the run}
}
\value{
the dataframe containing timing information
}
\description{
This function prints the parameter and timings information for a given run id.
Its main purpose is to explore a dataset.
}
\examples{
get_run_timings(conn, 29)
}
