% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timings_table.R
\name{get_set_timings}
\alias{get_set_timings}
\title{Get Latency of Benchmark Sets as Table}
\usage{
get_set_timings(df, unique_params, recipe_order = c(), group_results = T)
}
\arguments{
\item{df}{the dataframe prepared by the prepare_timings_data_comparison() function}

\item{unique_params}{the set of unique parameters that shall be returned as columns}

\item{recipe_order}{the order of the recipes within the table}

\item{group_results}{whether to aggregate the results by the unique params, defaults to `TRUE`}
}
\value{
a table of the latency values
}
\description{
This function takes a dataframe prepared by the prepare_timings_data_comparison() function
and returns the latency information formatted as a table.
}
\examples{
get_set_timings(
  df,
  unique_params = c("system", "recipe", "vectorize_type")
)
}
