% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_docker_stats.R
\name{parse_docker_stats}
\alias{parse_docker_stats}
\title{Parse Docker Stats DataFrame}
\usage{
parse_docker_stats(df, system_name, stage)
}
\arguments{
\item{df}{The dataframe containing splitted docker stats records}

\item{system_name}{the name of the system stats are requested for}

\item{stage}{the stage stats are requested for}
}
\value{
a dataframe containing the stats with only atomic values
}
\description{
This function converts the output of `docker stats` into atomic values.
}
\examples{
# assume a docker stats dataframe called `df` already exists
stats <- parse_docker_stats(df, "postgis", "ingestion")
}
