% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speedup.R
\name{plot_exec_speedup}
\alias{plot_exec_speedup}
\title{Create Execution Time Comparison}
\usage{
plot_exec_speedup(df, unique_params, get_plot = T)
}
\arguments{
\item{df}{the dataframe prepared by the prepare_timings_data_comparison() function}

\item{unique_params}{the set of unique parameters that shall be displayed on the x axis}

\item{get_plot}{wheter to return a plot or a dataframe}
}
\value{
a ggplot2 plot or a dataframe
}
\description{
This function was used in the past to easily compare cold starts and warm starts.
It has not been tested in a while and therefore may not work anymore
}
\examples{
plot_exec_speedup(
  df,
  c("system", "recipe")
)

plot_exec_speedup(
  df,
  c("system", "vectorize_type"),
  get_plot=F
)
}
