% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latency_stage.R
\name{plot_stage_comparison}
\alias{plot_stage_comparison}
\title{Plot Parameter Comparison for a Given Stage}
\usage{
plot_stage_comparison(
  df,
  selected_stage,
  unique_params,
  recipe_order = c(),
  facet_by = "",
  fill_by = "system",
  barplot = T,
  NA_x_height = 2,
  fill_by_palette = c()
)
}
\arguments{
\item{df}{the dataframe prepared by the prepare_timings_data_comparison() function}

\item{selected_stage}{the stage the graph shall be drawn for}

\item{unique_params}{the set of unique parameters that shall be displayed on the x axis}

\item{recipe_order}{the order of the recipes within the graph}

\item{facet_by}{the column (usually a parameter column) to facet the graph on,defaults to no facets}

\item{fill_by}{the column (usually a parameter column) determining the fill color of the bars or lines, defaults to `system`}

\item{barplot}{whether to create a bar plot or a line plot, defaults to `TRUE`}

\item{NA_x_height}{the y position of the "🗙" indicating a failed run, defaults to `2`}

\item{fill_by_palette}{the color palette the bars shall be filled with, defaults to the ggplot2 standard palette}
}
\value{
a ggplot2 plot
}
\description{
This is the main drawing function for ST_Benchi. It takes a dataframe prepared by the
prepare_timings_data_comparison() function and can plot a wide variety of comparisons
depending on the set of parameters supplied. Different to the `plot_full_run()` function,
this function only plots a single stage at once, at the benefit of more fine-granular settings
for this stage.
}
\details{
The function can group and facet the graph by different parameters.

If a run fails to complete (e.g. due to a timeout or error), the function will draw
an "🗙" near the x axis.
}
\examples{
library(RColorBrewer)
systems_colors <- setNames(brewer.pal(n=5, name="Accent"),
  c("beast", "heavyAI", "postgis", "rasdaman", "sedona"))

plot_stage_comparison(
  df,
  selected_stage = "execution",
  unique_params = c("system", "recipe"),
  facet_by = "vectorize_type",
  recipe_order = some_recipe_order_vector,
  fill_by = "system",
  fill_by_palette = systems_colors
)
}
