% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{prepare_timings_data_comparison}
\alias{prepare_timings_data_comparison}
\title{Load Run Sets from DB}
\usage{
prepare_timings_data_comparison(conn, ...)
}
\arguments{
\item{conn}{the database connection}

\item{...}{the set IDs to load}
}
\value{
the latency data asl wellas parameters for the different stages of a run
}
\description{
This is the main function for fetching latency data from the database. It uses
the benchmark set ids to identify relevant runs. If benchmarks are performed in
groups (e.g. using 2 raster and 3 vector datasets to perform 6 benchmark sets),
it is possible to load all six sets at the same time.
}
\details{
The results are returned in long form regarding the stage.

The duration returned is always the difference between the `start` and the `end` marker.
If a `pre` marker exists (as it occurs in some ingestion stages), this duration is
also returned and marked accordingly in the `ingest_dur_type` column. Similar, the
result also contains one row per cold/warm start.

Each row of the returned dataframe also contains the parameters of each run.
}
\examples{
conn <- dbConnect(duckdb::duckdb(), dbdir="results.db")
prepare_timings_data_comparison(conn, 23)
prepare_timings_data_comparison(conn, 17, 18, 19)
}
